Rebased with main and created HomeService with data mocked
