--- solutions/devsprint-eduardo-bocato-2/FinanceApp/Core/Services/HomeService.swift
@@ -1,12 +1,0 @@
-//
-//  HomeService.swift
-//  FinanceApp
-//
-//  Created by Gustavo Soares on 05/05/22.
-//
-
-import Foundation
-
-protocol HomeServiceProtocol {
-    static func fetchHomeData(_ completion: @escaping (HomeData?) -> Void)
-}
